Initial commit of Directional TripWire
A pair of laser beams aimed at photoresistors are placed in a series in
the direction of the motion to be tracked. Moving in one direction (i.e.
breaking beam 1 before beam 2) triggers an event. Moving in the opposite
direciton (i.e. breaking beam 2 before beam 1) performs a different
action (or no action).